+++ src/auto_flutter_dig/model/error/__init__.py
@@ -1,3 +1,4 @@
+from .chain import E
 from .error import *
 from .silent_warning import SilentWarning
